# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trentino Stage API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/profiles": {
      "get": {
        "summary": "Retrieve profile information",
        "description": "Given the profile ID and the authentication token, returns the profile information. If no profile ID is provided, it retrieves the user's own profile.",
        "tags": [
          "Profiles"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session token for authentication"
          },
          {
            "in": "query",
            "name": "profileId",
            "schema": {
              "type": "string"
            },
            "required": false,
            "description": "ID of the profile to retrieve (if not provided the profile of the user making the request is returned)"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "surname": {
                      "type": "string"
                    },
                    "bio": {
                      "type": "string"
                    },
                    "website": {
                      "type": "string"
                    },
                    "birth_date": {
                      "type": "string",
                      "format": "date"
                    },
                    "address": {
                      "type": "object",
                      "properties": {
                        "country": {
                          "type": "string"
                        },
                        "region": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string"
                        },
                        "postal_code": {
                          "type": "string"
                        },
                        "street": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        }
                      }
                    },
                    "identifier": {
                      "type": "string"
                    },
                    "sector": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "isCompany": {
                      "type": "boolean"
                    },
                    "profile_image": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile ID or other bad request"
          },
          "401": {
            "description": "Invalid session token"
          },
          "404": {
            "description": "Profile not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new profile",
        "description": "Creates a new profile for the authenticated user. The user must not already have a profile.",
        "tags": [
          "Profiles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "isCompany": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "surname": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "postal_code": {
                    "type": "string"
                  },
                  "street": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "birth_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "profile_image": {
                    "type": "string",
                    "format": "binary"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "identifier": {
                    "type": "string"
                  },
                  "sector": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile created successfully"
          },
          "400": {
            "description": "Invalid input or unsupported content type"
          },
          "401": {
            "description": "Invalid session token or user already has a profile"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Edit an existing profile",
        "description": "Updates the profile information for the authenticated user.",
        "tags": [
          "Profiles"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "sessionToken": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "surname": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "postal_code": {
                    "type": "string"
                  },
                  "street": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "sector": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile modified successfully"
          },
          "400": {
            "description": "Invalid input or unsupported content type"
          },
          "401": {
            "description": "Invalid session token"
          },
          "404": {
            "description": "Profile not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}